Initial commit
The “package-lock.json” file includes information on every single dependency within your app, even dependencies of other dependencies. It is like a snapshot of the current version of your entire package. If one of your dependencies were to change or be updated, your own package or project could break or cease to work. So long as you have the “package-lock.json” file, however, you have the blueprints to rebuild your package using dependencies that are guaranteed to work.