Initial commit
A “package.json” file is an essential part of a Node project, which can do a lot of things. It holds various metadata relevant to the project, lists any other packages your project depends on, and supplies the specific versions of each of your package’s dependencies. Among the many fields it contains, the only information that’s actually required in order to have a fully functioning “package.json” file is the version and the name.